move genUrl from plugins polio to iaso. Convert to typescript
