+++ hat/assets/js/apps/Iaso/routing/routing.ts
@@ -1,12 +1,6 @@
-import { formatPattern } from 'react-router';
-import { RouteConfig } from 'react-router-config';
-
-interface Router {
-  routes: RouteConfig[];
-  params: Record<string, string>;
-}
+import { formatPattern, Router } from 'react-router';
 
 const genUrl = (router: Router, newParams: Record<string, string>): string =>
-  formatPattern(router.routes[0].path, { ...router.params, ...newParams });
+    formatPattern(router.routes[0].path, { ...router.params, ...newParams });
 
 export { genUrl };
